IGNITE-28806 Refactor DynamicCacheStartExchangeTest
