+++ modules/core/src/test/java/org/apache/ignite/testsuites/IgniteCacheTestSuite6.java
@@ -21,10 +21,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import org.apache.ignite.cache.affinity.PendingExchangeTest;
+import org.apache.ignite.internal.processors.cache.CacheAffinityCoordinatorInitTest;
 import org.apache.ignite.internal.processors.cache.CacheIgniteOutOfMemoryExceptionTest;
 import org.apache.ignite.internal.processors.cache.CacheNoAffinityExchangeTest;
 import org.apache.ignite.internal.processors.cache.ClientFastReplyCoordinatorFailureTest;
-import org.apache.ignite.internal.processors.cache.DynamicCacheStartExchangeTest;
 import org.apache.ignite.internal.processors.cache.IgniteOutOfMemoryPropagationTest;
 import org.apache.ignite.internal.processors.cache.PartitionedAtomicCacheGetsDistributionTest;
 import org.apache.ignite.internal.processors.cache.PartitionedTransactionalPessimisticCacheGetsDistributionTest;
@@ -119,7 +119,7 @@ public class IgniteCacheTestSuite6 {
 
         GridTestUtils.addTestIfNeeded(suite, TransactionContextCleanupTest.class, ignoredTests);
 
-        GridTestUtils.addTestIfNeeded(suite, DynamicCacheStartExchangeTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, CacheAffinityCoordinatorInitTest.class, ignoredTests);
 
         return suite;
     }
